Fix #170 File format
This handles the case where column order might differ in the data files.
Instead always use the column order as defined by the schema and add
empty values whenever the columns are missing. One thing to note: we use
the component's dependencies to determine what columns should be
included. Not sure if this includes both required + optional columns,
occasionally there seem to be some columns in the manifest that are not
in the schema. The current logic excludes those.